+++ darwin/home.nix
@@ -4,6 +4,8 @@
 
   imports = [
     ./zsh.nix
+    ./kitty.nix
+    # ./alacritty.nix
   ];
 
   # Home Manager needs a bit of information about you and the
